upgrade common version to 2.4.7, refactor import for InstitutionType
