+++ connector-api/src/main/java/it/pagopa/selfcare/onboarding/connector/api/ProductsConnector.java
@@ -1,6 +1,6 @@
 package it.pagopa.selfcare.onboarding.connector.api;
 
-import it.pagopa.selfcare.onboarding.connector.model.onboarding.InstitutionType;
+import it.pagopa.selfcare.commons.base.utils.InstitutionType;
 import it.pagopa.selfcare.onboarding.connector.model.product.Product;
 
 public interface ProductsConnector {
