+++ connector-api/src/main/java/it/pagopa/selfcare/onboarding/connector/model/institutions/Institution.java
@@ -1,8 +1,8 @@
 package it.pagopa.selfcare.onboarding.connector.model.institutions;
 
+import it.pagopa.selfcare.commons.base.utils.InstitutionType;
 import it.pagopa.selfcare.onboarding.connector.model.onboarding.DataProtectionOfficer;
 import it.pagopa.selfcare.onboarding.connector.model.onboarding.GeographicTaxonomy;
-import it.pagopa.selfcare.onboarding.connector.model.onboarding.InstitutionType;
 import it.pagopa.selfcare.onboarding.connector.model.onboarding.PaymentServiceProvider;
 import lombok.Data;
