+++ connector-api/src/main/java/it/pagopa/selfcare/onboarding/connector/model/institutions/InstitutionInfo.java
@@ -1,8 +1,8 @@
 package it.pagopa.selfcare.onboarding.connector.model.institutions;
 
 import it.pagopa.selfcare.commons.base.security.PartyRole;
+import it.pagopa.selfcare.commons.base.utils.InstitutionType;
 import it.pagopa.selfcare.onboarding.connector.model.onboarding.Billing;
-import it.pagopa.selfcare.onboarding.connector.model.onboarding.InstitutionType;
 import lombok.Data;
 
 import java.util.Objects;
