--- connector-api/src/main/java/it/pagopa/selfcare/onboarding/connector/model/onboarding/InstitutionType.java
@@ -1,10 +1,0 @@
-package it.pagopa.selfcare.onboarding.connector.model.onboarding;
-
-public enum InstitutionType {
-    PA,
-    PG,
-    GSP,
-    PT,
-    SCP,
-    PSP
-}
