+++ connector-api/src/main/java/it/pagopa/selfcare/onboarding/connector/model/onboarding/OnboardingResponseData.java
@@ -1,6 +1,7 @@
 package it.pagopa.selfcare.onboarding.connector.model.onboarding;
 
 import it.pagopa.selfcare.commons.base.security.PartyRole;
+import it.pagopa.selfcare.commons.base.utils.InstitutionType;
 import it.pagopa.selfcare.onboarding.connector.model.RelationshipState;
 import it.pagopa.selfcare.onboarding.connector.model.institutions.Attribute;
 import lombok.Data;
