+++ connector-api/src/main/java/it/pagopa/selfcare/onboarding/connector/model/onboarding/PnPGOnboardingData.java
@@ -12,6 +12,7 @@
 
 package it.pagopa.selfcare.onboarding.connector.model.onboarding;
 
+import it.pagopa.selfcare.commons.base.utils.InstitutionType;
 import lombok.Data;
 import lombok.NoArgsConstructor;
