+++ connector/rest/src/main/java/it/pagopa/selfcare/onboarding/connector/rest/model/BillingDataResponse.java
@@ -1,7 +1,7 @@
 package it.pagopa.selfcare.onboarding.connector.rest.model;
 
+import it.pagopa.selfcare.commons.base.utils.InstitutionType;
 import it.pagopa.selfcare.onboarding.connector.model.onboarding.Billing;
-import it.pagopa.selfcare.onboarding.connector.model.onboarding.InstitutionType;
 import lombok.Data;
 
 @Data
