+++ core/src/main/java/it/pagopa/selfcare/onboarding/core/ProductService.java
@@ -1,6 +1,6 @@
 package it.pagopa.selfcare.onboarding.core;
 
-import it.pagopa.selfcare.onboarding.connector.model.onboarding.InstitutionType;
+import it.pagopa.selfcare.commons.base.utils.InstitutionType;
 import it.pagopa.selfcare.onboarding.connector.model.product.Product;
 
 public interface ProductService {
