+++ core/src/test/java/it/pagopa/selfcare/onboarding/core/ProductServiceImplTest.java
@@ -1,7 +1,7 @@
 package it.pagopa.selfcare.onboarding.core;
 
+import it.pagopa.selfcare.commons.base.utils.InstitutionType;
 import it.pagopa.selfcare.onboarding.connector.api.ProductsConnector;
-import it.pagopa.selfcare.onboarding.connector.model.onboarding.InstitutionType;
 import it.pagopa.selfcare.onboarding.connector.model.product.Product;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
