+++ pom.xml
@@ -22,7 +22,7 @@
     <description>Self Care OnBoarding Backend</description>
 
     <properties>
-        <selc-commons.version>2.4.6</selc-commons.version>
+        <selc-commons.version>2.4.7</selc-commons.version>
     </properties>
 
     <dependencyManagement>
