+++ web/src/main/java/it/pagopa/selfcare/onboarding/web/model/OnboardingDto.java
@@ -1,7 +1,7 @@
 package it.pagopa.selfcare.onboarding.web.model;
 
 import io.swagger.annotations.ApiModelProperty;
-import it.pagopa.selfcare.onboarding.connector.model.onboarding.InstitutionType;
+import it.pagopa.selfcare.commons.base.utils.InstitutionType;
 import lombok.Data;
 
 import javax.validation.Valid;
