+++ web/src/test/java/it/pagopa/selfcare/onboarding/web/model/CompanyOnboardingDtoTest.java
@@ -1,6 +1,6 @@
 package it.pagopa.selfcare.onboarding.web.model;
 
-import it.pagopa.selfcare.onboarding.connector.model.onboarding.InstitutionType;
+import it.pagopa.selfcare.commons.base.utils.InstitutionType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
